Stream video from drone to <video> tag
There is still a barely perceptible delay but it's less than 200ms which
is better than any other approach I've tried, including the official
android app.

JMuxer seems to expect an SPS and PPS NAL in each set of NALs passed
to `feed`. The drone only sends these every second or so, which gives us
a lot of unnecessary lag. As a workaround, `TelloControl.Video.Processor`
saves the most recently seen SPS and PPS NALs and attaches them before
every video NAL. This adds about 20 bytes per 15kB, so I'm not too
worried about the overhead.

I set the frame rate in JMuxer to 40 fps even though the actual video
feed is about 30 fps so that it never buffers.